+++ changelog.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 0.1.3
+- Module actions/mutations/getters can now be non-namespaced by supplying `namespaced : false` in the configuration.
+- All modules are namespaced by default, but this can be switched off by setting `mockVuex.config.autoNamespace = false`
+
 ## 0.1.2
 - Added missing `rootGetters` argument from the getters method [#4](https://github.com/jackmellis/mock-vuex/issues/4)
 - Creating a store with an action and mutation that share the same name did not work. The action would overwrite the mutation [#3](https://github.com/jackmellis/mock-vuex/issues/3)